Split files in HIP backend
